+++ decorators/decorator.rb
@@ -1,12 +1,12 @@
 require_relative './nameable'
 
 class Decorator < Nameable
- 
-    def initialize(nameable = Nameable.new())
-        super()
-        @nameable = nameable
-    end 
-    def correct_name
-        @nameable.correct_name
-    end 
+  def initialize(nameable = Nameable.new())
+    super()
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
 end
