+++ decorators/nameable.rb
@@ -1,5 +1,5 @@
-class Nameable 
-    def correct_name
-        raise NotImplementedError
-    end 
-end 
+class Nameable
+  def correct_name
+    raise NotImplementedError
+  end
+end
