+++ decorators/trimmer_decorator.rb
@@ -1,0 +1,7 @@
+require_relative './decorator'
+
+class TrimmerDecorator
+  def correct_name
+    @nameable.correct_name[0..9] if @nameable.correct_name.length > 10
+  end
+end
