+++ entities/person.rb
@@ -2,6 +2,7 @@ require_relative './decorators/nameable'
 
 class Person < Nameable
   def initialize(age, name = 'unknown', parent_permission: true)
+    super()
     @id = Random.rand(1..1000)
     @age = age
     @parent_permission = parent_permission
@@ -21,6 +22,5 @@ class Person < Nameable
 
   def correct_name
     @name
-  end 
-
+  end
 end
